refactor: reword some comments in workflow module, add Command typedef
