+++ src/core/workflow/index.js
@@ -11,9 +11,98 @@ import {
   emptyReporter, JobStatuses, ResultDetailTypes,
 } from '@core/workflow/reporter';
 
-export * from '@core/workflow/config';
+export * from '@core/workflow/command';
 export * from '@core/workflow/reporter';
 export { default as createStepper } from '@core/workflow/stepper';
+
+/**
+ * @typedef {import('@core/workflow/reporter').Reporter} Reporter
+ * @typedef {import('@core/workflow/reporter').Progress} Progress
+ */
+
+/**
+ * @typedef {object} Action
+ *
+ * `Action` is a usable command or function which can be used in combination with a set of
+ * arguments to create `Job`(s).
+ *
+ * @property {(object) => any} validate The action's validation logic.
+ * @property {(object, object) => any} run The action's execution logic.
+ */
+
+/**
+ * Create a usable `Action`.
+ *
+ * @param {(object) => any} validate The action's `validate` function.
+ * @param {(object, object) => any} run The action's `run` function.
+ * @returns {Action} The `Action`.
+ */
+export function createAction(validate, run) {
+  return { validate, run };
+}
+
+/**
+ * @typedef {object} Job
+ *
+ * `Job` is a part of a `Workflow`. A `Workflow` is a sequential list of `Job`(s).
+ *
+ * @property {string} name The job's name.
+ * @property {() => any} validate The job's validation logic.
+ * @property {(object) => any} run The job's execution logic.
+ */
+
+/**
+ * Create a `Job`.
+ *
+ * @param {string} name The job's name.
+ * @param {() => any} validate The job's `validate` function.
+ * @param {(object) => any} run The job's `run` function.
+ * @param {object} args The job's arguments.
+ * @returns {Job} The `Job`.
+ */
+export function createJob(name, validate, run, args) {
+  return {
+    name,
+    validate: validate ? () => validate(args) : undefined,
+    run: (opts = {}) => run(args, opts),
+  };
+}
+
+/**
+ * @typedef {object} JobConfig The `Job` configuration model.
+ *
+ * @property {string} uses The action to use for the job.
+ * @property {object} args The job's arguments.
+ */
+
+/**
+ * @typedef {object} WorkflowConfig The `Workflow` configuration model.
+ *
+ * @property {JobConfig[]} jobs The `Workflow`'s `Job` configuration objects.
+ */
+
+/**
+ * Load `Job`(s) for a `Workflow` based on its configuration object.
+ *
+ * @param {object} library The library of usable `Action`(s).
+ * @param {WorkflowConfig} config The `Workflow` configuration object.
+ * @returns {Job[]} The `Workflow`'s `Job`(s).
+ */
+export function loadJobs(library, config) {
+  const jobs = [];
+
+  config.jobs.forEach((jobConfig) => {
+    const { uses, args } = jobConfig;
+    if (!has(library, uses)) {
+      throw new Error(`unknown action '${uses}'`);
+    }
+
+    const action = library[uses];
+    jobs.push(createJob(uses, action.validate, action.run, args ?? {}));
+  });
+
+  return jobs;
+}
 
 // Some texts.
 const Strings = {
@@ -22,20 +111,14 @@ const Strings = {
 };
 
 /**
- * @typedef {import('@core/workflow/config').Job} Job
- * @typedef {import('@core/workflow/reporter').Reporter} Reporter
- * @typedef {import('@core/workflow/stepper').Progress} Progress
- */
-
-/**
  * Run a `Workflow`, which is essentially a sequential list of `Job`(s).
  *
  * @param {Job[]} jobs The `Workflow`'s `Job`(s).
  * @param {object} opts Options.
- * @param {Reporter} opts.reporter A `Reporter` object for the `Workflow` to report progress and
- * result.
+ * @param {Reporter} opts.reporter A `Reporter` object for the `Workflow` to report its progress
+ * and result.
  * @returns {boolean} Whether the `Workflow` succeeded or failed. The `Workflow` is considered
- * succeeded only if all of its Job(s) succeeded, otherwise the `Workflow` is failed.
+ * succeeded only if all of its Job(s) succeeded, otherwise the `Workflow` failed.
  */
 export default function runWorkflow(jobs, opts = {}) {
   const { reporter } = { ...runWorkflow.defaultOpts, ...opts };
